Traffic-main/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Database
influxdb-client>=1.38.0
redis>=5.0.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0

# Data Processing
kafka-python>=2.0.0
# pyspark>=3.5.0

# Web and UI
streamlit>=1.29.0
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# IoT and Sensors
paho-mqtt>=1.6.0
# opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
asyncio-mqtt>=0.16.0
# kaggle>=1.6.17
